Fix a couple of issues that regress converter performance
AbstractKafkaAvroDeserializer copies the schema if the subject name strategy
implements the deprecated interface. All the available strategies implement
both the deprecated and new interfaces, and therefore the schema is copied
on every call to toConnectData, which is very expensive. This patch changes
the deprecation check to return true if the strategy does not implement the
new interface.

AbstractKafkaAvroDeserializer always sets the version property when used by
the converter. This resets the hash code inside the avro schema, which is
completely unnecessary in most cases where the property has already been set.
Computing the hash code for the schema can be expensive, so we want to avoid
doing so. To prevent this check whether the version is set before calling
addProp.